Refactor single line grammar node package
