+++ src/main/java/com/github/freenamu/parser/grammar/CompositeGrammar.java
@@ -1,7 +1,7 @@
 package com.github.freenamu.parser.grammar;
 
 import com.github.freenamu.parser.node.Node;
-import com.github.freenamu.parser.node.Text;
+import com.github.freenamu.parser.node.single.Text;
 
 import java.util.ArrayList;
 import java.util.List;
