+++ src/main/java/com/github/freenamu/parser/grammar/singleline/FootnoteGrammar.java
@@ -1,9 +1,9 @@
 package com.github.freenamu.parser.grammar.singleline;
 
 import com.github.freenamu.parser.grammar.LeafGrammar;
-import com.github.freenamu.parser.node.Footnote;
 import com.github.freenamu.parser.node.Node;
-import com.github.freenamu.parser.node.Text;
+import com.github.freenamu.parser.node.single.Footnote;
+import com.github.freenamu.parser.node.single.Text;
 
 import java.util.ArrayList;
 import java.util.List;
