+++ src/main/java/com/github/freenamu/parser/node/single/Strikeout.java
@@ -1,4 +1,6 @@
-package com.github.freenamu.parser.node;
+package com.github.freenamu.parser.node.single;
+
+import com.github.freenamu.parser.node.Node;
 
 import java.util.List;
