+++ src/main/java/com/github/freenamu/parser/node/single/Text.java
@@ -1,4 +1,6 @@
-package com.github.freenamu.parser.node;
+package com.github.freenamu.parser.node.single;
+
+import com.github.freenamu.parser.node.Node;
 
 public class Text extends Node {
     private final String text;
